Move `tcp_listener` to a dedicated file and reduce header coupling
