+++ nano/core_test/socket.cpp
@@ -2,6 +2,7 @@
 #include <nano/boost/asio/ip/network_v6.hpp>
 #include <nano/lib/thread_runner.hpp>
 #include <nano/node/transport/socket.hpp>
+#include <nano/node/transport/tcp_listener.hpp>
 #include <nano/test_common/system.hpp>
 #include <nano/test_common/testutil.hpp>
