+++ nano/node/CMakeLists.txt
@@ -158,6 +158,8 @@ add_library(
   transport/socket.cpp
   transport/tcp.hpp
   transport/tcp.cpp
+  transport/tcp_listener.hpp
+  transport/tcp_listener.cpp
   transport/tcp_server.hpp
   transport/tcp_server.cpp
   transport/transport.hpp
